refactor: 폴더 구조 수정(페이지 별 컴포넌트 폴더 생성)
