+++ src/pages/Library/Library.tsx
@@ -2,13 +2,13 @@ import React, { useEffect, useState } from 'react'
 import { Pagination } from '@mui/material'
 import { Link } from 'react-router-dom'
 import { useQuery } from '@tanstack/react-query'
-import Header from '../layouts/Header'
-import Footer from '../layouts/Footer'
-import Wrapper from '../layouts/Wrapper'
-import { getLibrary, LibraryType } from '../services/library'
-import { Status, statusEn } from '../types/bookType'
-import getFormattedIsbn from '../utils/getFormattedIsbn'
-import getSearchBook from '../services/searchBook'
+import Header from '../../components/Header'
+import Footer from '../../components/Footer'
+import Wrapper from '../../components/Wrapper'
+import { getLibrary, LibraryType } from '../../services/library'
+import { Status, statusEn } from '../../types/bookType'
+import getFormattedIsbn from '../../utils/getFormattedIsbn'
+import getSearchBook from '../../services/searchBook'
 
 interface LibraryListType {
     isbn: string
