+++ src/pages/LibraryDetail/LibraryDetail.tsx
@@ -3,25 +3,25 @@ import { Chip, Pagination } from '@mui/material'
 import { redirect, useParams } from 'react-router-dom'
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query'
 import { AxiosError } from 'axios'
-import Header from '../layouts/Header'
-import Footer from '../layouts/Footer'
-import Wrapper from '../layouts/Wrapper'
-import SearchItem from '../components/SearchItem'
-import ReadRecordInfo from '../components/ReadRecordInfo'
-import Record from '../components/Record'
-import ReadingRecordInfo from '../components/ReadingRecordInfo'
-import WishRecordInfo from '../components/WishRecordInfo'
-import { statusKo } from '../types/bookType'
-import { searchDocumentType } from '../types/searchResultType'
-import getSearchBook from '../services/searchBook'
+import Header from '../../components/Header'
+import Footer from '../../components/Footer'
+import Wrapper from '../../components/Wrapper'
+import SearchItem from '../Search/components/SearchItem'
+import ReadRecordInfo from './components/Record/ReadRecordInfo'
+import Record from './components/Record/Record'
+import ReadingRecordInfo from './components/Record/ReadingRecordInfo'
+import WishRecordInfo from './components/Record/WishRecordInfo'
+import { statusKo } from '../../types/bookType'
+import { searchDocumentType } from '../../types/searchResultType'
+import getSearchBook from '../../services/searchBook'
 import {
     deleteLibrary,
     getLibraryDetail,
     getRecordList,
-} from '../services/library'
-import { LibraryDetailType, RecordListType } from '../types/libraryType'
-import RecordingModal from '../components/RecordingModal'
-import UpdateBookModal from '../components/UpdateBookModal'
+} from '../../services/library'
+import { LibraryDetailType, RecordListType } from '../../types/libraryType'
+import RecordingModal from './components/Record/RecordingModal'
+import UpdateBookModal from './components/UpdateBookModal'
 
 export default function LibraryDetail() {
     const { isbn } = useParams()
