+++ src/pages/LibraryDetail/components/Record/ReadingRecordInfo.tsx
