+++ src/pages/LibraryDetail/components/Record/WishRecordInfo.tsx
