+++ src/pages/Search/components/SearchItem.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
-import getFormattedDate from '../utils/getFormattedDate'
-import getFormattedIsbn from '../utils/getFormattedIsbn'
-import { searchDocumentType } from '../types/searchResultType'
+import getFormattedDate from '../../../utils/getFormattedDate'
+import getFormattedIsbn from '../../../utils/getFormattedIsbn'
+import { searchDocumentType } from '../../../types/searchResultType'
 
 export default function SearchItem({
     searchResult,
